glm-47-finetune/requirements.txt
torch>=2.0.0
unsloth[z奖]>=2025.1.15
transformers>=4.46.0
peft>=0.13.0
trl>=0.14.0
datasets>=3.0.0
accelerate>=0.34.0
bitsandbytes>=0.44.0
wandb>=0.18.0
scikit-learn>=1.5.0
scipy>=1.14.0
tqdm>=4.66.0
pyyaml>=6.0.0
